[MICRSE-1666] Modified wayland.c and lv_conf.h to use two draw buffers and the full refresh feature.
This change solves the image overlaping issue, which was produced by the partial refresh feature
which is set by default unless full refresh is specified; also added the use of two draw buffers
to prevent race conditions in the draw buffer. As it can be tried to access to it simultaneously
by the read/write functions that share this resource.

Signed-off-by: Ohtokani Pinzon <ohtokani.pinzon@nxp.com>